requirements.txt
autopep8==1.5.5
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycodestyle==2.6.0
PyYAML==5.4.1
toml==0.10.2
Werkzeug==1.0.1
numpy==1.20.1
opencv-python==4.5.1.48
numpy>1.0,<1.15